+++ CHANGES.md
@@ -6,6 +6,7 @@
 
 ### Merged from 1.0
 
+* Fix logging fault reporter raising duplicate alarm - Issue #557
 * Fix priority calculation for local queue - Issue #546
 * Skip unnecessary reads from repair history - Issue #548
 * Fix repair job priority - Issue #515
@@ -18,6 +19,7 @@
 
 #### Merged from 1.0
 
+* Fix logging fault reporter raising duplicate alarm - Issue #557
 * Fix priority calculation for local queue - Issue #546
 * Skip unnecessary reads from repair history - Issue #548
 * Fix repair job priority - Issue #515
@@ -57,6 +59,7 @@
 
 ## Version 1.0.8 (Not yet released)
 
+* Fix logging fault reporter raising duplicate alarm - Issue #557
 * Fix priority calculation for local queue - Issue #546
 * Skip unnecessary reads from repair history - Issue #548
 * Fix repair job priority - Issue #515