+++ CHANGES.md
@@ -2,6 +2,10 @@
 
 ## Version 4.0.6 (Not yet released)
 
+### Merged from 1.0
+
+* Fix logging fault reporter raising duplicate alarm - Issue #557
+
 ## Version 4.0.5
 
 ### Merged from 1.0
@@ -61,6 +65,7 @@
 
 ### Merged from 1.0
 
+* Fix logging fault reporter raising duplicate alarm - Issue #557
 * Fix priority calculation for local queue - Issue #546
 * Skip unnecessary reads from repair history - Issue #548
 * Fix repair job priority - Issue #515
@@ -83,6 +88,7 @@
 
 ### Merged from 1.0
 
+* Fix logging fault reporter raising duplicate alarm - Issue #557
 * Fix priority calculation for local queue - Issue #546
 * Skip unnecessary reads from repair history - Issue #548
 * Fix repair job priority - Issue #515
@@ -207,6 +213,7 @@
 
 ### Merged from 1.0
 
+* Fix logging fault reporter raising duplicate alarm - Issue #557
 * Fix priority calculation for local queue - Issue #546
 * Skip unnecessary reads from repair history - Issue #548
 * Fix repair job priority - Issue #515
@@ -219,6 +226,7 @@
 
 #### Merged from 1.0
 
+* Fix logging fault reporter raising duplicate alarm - Issue #557
 * Fix priority calculation for local queue - Issue #546
 * Skip unnecessary reads from repair history - Issue #548
 * Fix repair job priority - Issue #515
@@ -258,6 +266,7 @@
 
 ## Version 1.0.8 (Not yet released)
 
+* Fix logging fault reporter raising duplicate alarm - Issue #557
 * Fix priority calculation for local queue - Issue #546
 * Skip unnecessary reads from repair history - Issue #548
 * Fix repair job priority - Issue #515